Updated authentication-setting service
